security: strengthen API route parameter validation and sanitization
