+++ src/lib/sanitizer.ts
@@ -1,0 +1,23 @@
+export function stripHtml(str: string): string {
+  return str.replace(/[<>&"']/g, (c) => {
+    switch (c) {
+      case "<":
+        return "&lt;";
+      case ">":
+        return "&gt;";
+      case "&":
+        return "&amp;";
+      case '"':
+        return "&quot;";
+      case "'":
+        return "&#x27;";
+      default:
+        return c;
+    }
+  });
+}
+
+export function sanitizeString(value: unknown, maxLength = 500): string {
+  if (typeof value !== "string") return "";
+  return stripHtml(value.trim().slice(0, maxLength));
+}
